requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
black==25.1.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
filelock==3.18.0
firebase-admin==6.8.0
flake8==7.2.0
fonttools==4.58.0
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jiter==0.10.0
joblib==1.5.1
jsonpickle==4.1.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.7.0
msgpack==1.1.0
music21==9.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==1.26.4
openai==1.82.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.31.2
webcolors==24.11.1

